TRD tracking QC info is a proper DPL message
